requirements.txt
Django==2.2
Babel==2.6.0

dwebsocket==0.5.11

djangorestframework==3.9.2
djangorestframework-datatables
django-datatable-view-compat

django-allauth==0.39.1
django-countries==5.3.3
django-phonenumber-field==2.3.1
phonenumberslite==8.10.10

xlrd
xlwt

pip==19.1
setuptools==41.0.1
wheel==0.33.1

Faker==1.0.5
psycopg2>=2.7,<3.0

django-model-utils==1.1.0

pillow

django-extensions
django-guardian
